steward: unify execution engines into single path (Issue #412)
Controller-connected stewards previously used config/executor.go with only 3
modules (file, directory, script) and a simple Parse→Set workflow. Standalone
stewards used execution/execution.go with all 7 modules and the full
Get→Compare→Set→Verify workflow. This created a capability gap where
controller-connected stewards were less capable than standalone ones.

This change eliminates that gap by making both modes use the same underlying
ExecutionEngine:

- Add execution/executor.go: Executor wraps ExecutionEngine to provide the
  controller-mode interface (raw bytes in, ConfigStatusReport out) while using
  the unified engine with all 7 built-in modules (file, directory, script,
  firewall, package, patch, acme) and the Get→Compare→Set→Verify workflow.

- Remove config/executor.go: The 3-module executor is replaced by the new
  execution.Executor. Types used only by the executor (Config, ConfigurationSpec,
  ResourceSpec, genericConfigState, normalizeValue) are removed; validatePermissions
  is preserved in config/validation.go for continued test coverage.

- Update client/client_transport.go: InitializeConfigExecutor now instantiates
  execution.Executor instead of config.Executor, giving controller-connected
  stewards access to all 7 modules and idempotent drift detection.

- Add execution/executor_test.go: Tests verify all 7 modules are available,
  the Get→Compare→Set→Verify workflow is idempotent (second apply is no-op),
  error isolation between modules, and YAML/JSON parsing.

Fixes #412
